Lab5: Formatting
Joseph Schmidt, 10/11/2014, Lab5. I finished the lab yesterday but I
forgot to sync it until earlier today, and I also remembered that the
formatting gets weird when you edit the code in the pgAdmin editor, so I
fixed that (hope that's all okay in terms of not loosing points on the
lab).