Lab6: Finished ( I think)
Joseph Schmidt, 10/14/14, Lab6. It's amazing how such a precise art can
be so hard to judge in terms progress sometimes. Technically, I'm done
with this lab, but I could figure out tomorrow that I have a couple more
hours of work to put into this one. Testing, and other checks for the
requirements of this lab, are to follow in the next couple of days.